Moving UserFrosting, Facade & SlimAppEvent
